Fix logout consistency and workspace lock acquisition safety
